Add files via upload
It is an ML Based Recommender System which follows Content based recommendations that recommends movies based on the user's interest.
Suppose user 1 and user 2 has similar interests and user 1 watches a movie which he liked. so the recommender system will suggest the same movie to user 2 ,this is how this system works.
It calculates the similarity score which is a numerical value ranges between zero to one which helps to determine how much two items are similar to each other on a scale.of zero to one. This similarity score is obtained measuring the similarity between the text details of both of the items.
Text as in genre of the movie, cast of the movie, ratings of the movie etc